Explanation of how to break down this exercise
Using IRB we get nil for the puts line...
But that is because puts always returns nil.

Also, the evaluation of the line is separate from any assignments that
happen.  So we can both take advantage of this, as we may see in a
future week, but also we should be aware that every line in Ruby returns
something.